# Use `unlist()` in `join_rows()` until `vec_c()` performance improves