--- a/R/join-rows.R
+++ b/R/join-rows.R
@@ -14,7 +14,7 @@ join_rows <- function(x_key, y_key, type = c("inner", "left", "right", "full"), 
 
   # flatten index list
   x_loc <- rep(x_loc, lengths(y_loc))
-  y_loc <- vec_c(!!!y_loc, .ptype = integer())
+  y_loc <- index_flatten(y_loc)
 
   y_extra <- integer()
 
@@ -27,4 +27,10 @@ join_rows <- function(x_key, y_key, type = c("inner", "left", "right", "full"), 
   }
 
   list(x = x_loc, y = y_loc, y_extra = y_extra)
+}
+
+# TODO: Replace with `vec_unchop(x, ptype = integer())`
+# once performance of `vec_c()` matches `unlist()`. See #4964.
+index_flatten <- function(x) {
+  unlist(x, recursive = FALSE, use.names = FALSE)
 }